pyproject.toml
[project]
name = "teamstory"
version = "0.1.0"

description = "Team Story Announcement Board"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "litellm"
]

[project.scripts]
teamstory = "main:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.scripts]
